+++ Keyboard/Controllers/KeyboardViewController+CandidateDataSource.swift
@@ -69,7 +69,7 @@ extension KeyboardViewController: UICollectionViewDataSource, UICollectionViewDe
     }
 
     private func commitCandidate(_ item: CandidateItem) {
-        emitKeyPressFeedback()
+        emitFeedback(for: .commit)
         let effects: KeyboardEffect
         if let correction = item.correction {
             effects = controller.handle(.insertCorrectionCandidate(correction))